pyproject.toml
[tool.poetry]
name = "outomaatti"
version = "1.0.0"
description = "Outomaatti"
authors = ["Markus <markus.aurala@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">= 3.8, < 3.12"
invoke = "^2.2.0"
numpy = "1.24"
dynaconf = "^3.2.5"
pygame-ce = "^2.4.1"
pygame-menu-ce = "^4.4.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
coverage = "^7.4.4"
pylint = "^3.1.0"
autopep8 = "^2.1.0"
pyclean = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
